pyproject.toml
[tool.poetry]
name = "edge-assignment-temp"
version = "0.1.0"
description = ""
authors = ["Niccolò Salvini <niccolo.salvini27@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.8"
requests = "^2.28.1"
numpy = "^1.18.0"
opencv-python = "^3.4.13.47"
Pillow = "^9.2.0"
slackclient = "^2.9.4"
python-dotenv = "^0.21.0"
pydantic = "^1.10.2"

[tool.poetry.dev-dependencies]
pytest = "^7.1.3"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
